requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cajarename==19.7.15
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
deja-dup-caja==0.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
folder-color-caja==0.0.86
folder-color-common==0.0.86
future==0.18.2
gattlib==0.20201113
greenlet==1.0.0
httplib2==0.14.0
idna==2.8
keyring==18.0.1
keyrings.alt==3.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
Magnus==1.0.3
Mako==1.1.0
MarkupSafe==1.1.0
mate-hud==19.10.0
mate-menu==20.4.1
mate-tweak==20.4.0
monotonic==1.5
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
paho-mqtt==1.5.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
psutil==5.5.1
pulsemixer==1.5.0
PyBluez==0.23
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
PyNaCl==1.3.0
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-xapp==1.8.1
python-xlib==0.23
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.4.5
ssh-import-id==5.10
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
